refactor: localStorage model removed to separate dir
